NISTTests

Documentation for NISTTests.

NISTTests.R_summary_noint

NISTTests.compare_nist_beta

NISTTests.compare_nist_rsd

NISTTests.compare_nist_rsquare

NISTTests.compare_nist_standarderror

NISTTests.full_compare_noint

NISTTests.full_noint

NISTTests.get_nist_beta

NISTTests.get_nist_data

NISTTests.get_nist_rsd

NISTTests.get_nist_rsquare

NISTTests.get_nist_standarderror

NISTTests.how_same

NISTTests.julia_summary_filip

NISTTests.julia_summary_noint

NISTTests.R_summary_noint — Method

R_summary_noint

Uses R to calculate datas for NoInt1 and NoInt2 Datasets.

NISTTests.compare_nist_beta — Method

compare_nist_beta

Uses how_same to compare i-th beta from NIST-Data with input.

NISTTests.compare_nist_rsd — Method

compare_nist_rsd

Uses how_same to compare rsd from NIST-Data with input.

NISTTests.compare_nist_rsquare — Method

compare_nist_rsquare

Uses how_same to compare rsquare from NIST-Data with input.

NISTTests.compare_nist_standarderror — Method

compare_nist_standarderror

Uses how_same to compare i-th Standarderror from NIST-Data with input.

NISTTests.full_compare_noint — Method

full_compare_noint

Uses julia and R to calculate datas for NoInt1 and NoInt2 Datasets and compares them with actual NIST-Datasets and display.

NISTTests.full_noint — Method

full_noint

Uses julia and R to calculate datas for NoInt1 and NoInt2 Datasets and displays them with actual NIST-Datasets.

NISTTests.get_nist_beta — Method

get_nist_beta

Takes i and loads i-th Beta value for the specified NIST Dataset.

NISTTests.get_nist_data — Method

get_nist_data

Function to load y,x1,x2,.. Data from pecified NIST Dataset as a Dataframe Object.

NISTTests.get_nist_rsd — Method

get_nist_rsd

Returns the RSD value for the specified NIST Dataset.

NISTTests.get_nist_rsquare — Method

get_nist_rsquare

Return the R-Square value for the specified NIST Dataset.

NISTTests.get_nist_standarderror — Method

get_nist_standarderror

Takes i and loads i-th Standard error value for the specified NIST Dataset.

NISTTests.how_same — Method

how_same

A function to check how close two numbers are.

NISTTests.julia_summary_filip — Method

julia_summary_filip

Uses julia to calculate datas for Filip.

NISTTests.julia_summary_noint — Method

julia_summary_noint

Uses julia to calculate datas for NoInt1 and NoInt2 Datasets.